+++ data/FeatureCache.cc
@@ -103,7 +103,7 @@ bool   FeatureCache::update()
 
 void   FeatureCache::dump() const
 {
-  printf("FeatureCache entries %d\n",_names.size());
+  printf("FeatureCache entries %d\n",(int) _names.size());
   for(unsigned k=0; k<_names.size(); k++) {
     printf("  %s\n",_names[k].c_str());
   }
