+++ qt/TdcClient.cc
@@ -289,7 +289,7 @@ int  TdcClient::configure       (iovec* iov)
     (*it)->configure(p,_input_signature,_output_signature);
 
   if (p > _request+BufferSize) {
-    printf("Client request overflow: size = 0x%x\n", p-_request);
+    printf("Client request overflow: size = 0x%x\n", (unsigned) (p-_request));
     return 0;
   }
   else {
